refactor: 비밀번호 상수 파일을 constants 디렉토리로 이동
- PASSWORD_LENGTH, PasswordBoxStatus 등을 constants/password.ts로 이동
- 관련 컴포넌트 및 화면에서 해당 경로를 참조하도록 수정
