+++ frontend/src/constants/password.ts
